update demo with export/import/recovery tss key, refactor and fixed MFA
